refactor subagent registry into cohesive modules
